Ultima Actualización
CRUD completo y responsivo

+++ pages/EditarInventarios/[id]/index.jsx
@@ -9,6 +9,7 @@ const {query} = useRouter();
 const {id} = query;
 const [updatedObjects, setUpdatedObjects] = useState(props.objects);
 const [updatedObjectsP, setUpdatedObjectsP] = useState(props.objectsp);
+const [editingRowId, setEditingRowId] = useState(null);
 
 const handleFormSubmit = async (event) => {
   event.preventDefault(); 
@@ -27,7 +28,7 @@ const handleFormSubmit = async (event) => {
 
   try {
 
-    const response = await axios.post(`http://localhost:3000/api/objetos/${id}`, formData);
+    await axios.post(`http://localhost:3000/api/objetos/${id}`, formData);
     event.target.reset();
     const object = await axios(`http://localhost:3000/api/objetos/${id}`);
     setUpdatedObjects(object.data);
@@ -50,7 +51,7 @@ const handleFormSubmitP = async (event) => {
     return;
   }
   try {
-    const response = await axios.post(
+    await axios.post(
       `http://localhost:3000/api/objetos_p/${id}`,
       {
         objeto_p: nombreObjetoP,
@@ -69,19 +70,63 @@ const handleFormSubmitP = async (event) => {
 };
 
 const handlerDeleteObjects = async(idObjeto, tipoInventario)=>{
-  try{
-    if(Number(tipoInventario) === 0){
-      await axios.delete(`http://localhost:3000/api/deleteIndObjects/${Number(idObjeto)}`);
+  const confirmed = window.confirm("¿Estás seguro de que quieres eliminar este objeto?");
+  if (!confirmed) {
+    return;
+  }else{
+    try{
+      if(Number(tipoInventario) === 0){
+        await axios.delete(`http://localhost:3000/api/deleteIndObjects/${Number(idObjeto)}`);
+        const object = await axios(`http://localhost:3000/api/objetos/${id}`);
+        setUpdatedObjects(object.data);
+
+      }else if(Number(tipoInventario) === 1){
+        await axios.delete(`http://localhost:3000/api/deleteIndObjectsP/${Number(idObjeto)}`);
+        const objectsp = await axios(`http://localhost:3000/api/objetos_p/${id}`);
+        setUpdatedObjectsP(objectsp.data);
+      }
+    }catch(error){
+      console.log(error);
+    }
+  }
+};
+const handleUpdate = async (idObjeto,tipo) => {
+
+  if(tipo === 0){
+    const objeto = document.getElementById(`objeto-${idObjeto}`).value;
+    const tipo = document.getElementById(`tipo-${idObjeto}`).value;
+    const peso = document.getElementById(`peso-${idObjeto}`).value;
+    const valor = Number(document.getElementById(`valor-${idObjeto}`).value);
+    const cantidad = Number(document.getElementById(`cantidad-${idObjeto}`).value);
+
+    const formData = {objeto, tipo, peso, valor, cantidad,};
+
+    try {
+      await axios.post(`http://localhost:3000/api/ActIndObject/${idObjeto}`, formData);
+
       const object = await axios(`http://localhost:3000/api/objetos/${id}`);
       setUpdatedObjects(object.data);
-      
-    }else if(Number(tipoInventario) === 1){
-      await axios.delete(`http://localhost:3000/api/deleteIndObjectsP/${Number(idObjeto)}`);
-      const objectsp = await axios(`http://localhost:3000/api/objetos_p/${id}`);
-      setUpdatedObjectsP(objectsp.data);
+      setEditingRowId(null);
+    } catch (error) {
+      console.error(error);
     }
-  }catch(error){
-    console.log(error);
+  }else if(tipo === 1){
+    const objeto_p = document.getElementById(`objetoP-${idObjeto}`).value;
+  const reservacion = document.getElementById(`reservacion-${idObjeto}`).value;
+  const valor_p = document.getElementById(`valorP-${idObjeto}`).value;
+
+  const formDataP = {objeto_p, reservacion, valor_p};
+
+  try {
+    await axios.post(`http://localhost:3000/api/ActIndObjectP/${idObjeto}`, formDataP)
+    console.log(formDataP)
+    
+    const object = await axios(`http://localhost:3000/api/objetos_p/${id}`);
+    setUpdatedObjectsP(object.data);
+    setEditingRowId(null);
+  } catch (error) {
+    console.error(error);
+  }
   }
 };
     return(
@@ -106,16 +151,53 @@ const handlerDeleteObjects = async(idObjeto, tipoInventario)=>{
               <tbody>
               {updatedObjects.map((row, index) => (
                 <tr key={index}>
-                    <td>{row.id_objeto}</td>
-                    <td>{row.objeto}</td>
-                    <td>{row.tipo}</td>
-                    <td>{row.peso}</td>
-                    <td>{row.valor}</td>
-                    <td>{row.cantidad}</td>
-                    <td>{id}</td>
-                    <td><button id="boton3" onClick={() =>handlerDeleteObjects(row.id_objeto,props.tabla.tipo_inventario)}>Eliminar</button></td>
-                </tr>
-              ))}
+                <td>{row.id_objeto}</td>
+                <td>
+                  {editingRowId === row.id_objeto ? (
+                    <input type="text" defaultValue={row.objeto} id={`objeto-${row.id_objeto}`} />
+                  ) : (
+                    row.objeto
+                  )}
+                </td>
+                <td>
+                  {editingRowId === row.id_objeto ? (
+                    <input type="text" defaultValue={row.tipo} id={`tipo-${row.id_objeto}`} />
+                  ) : (
+                    row.tipo
+                  )}
+                </td>
+                <td>
+                  {editingRowId === row.id_objeto ? (
+                    <input type="text" defaultValue={row.peso} id={`peso-${row.id_objeto}`} />
+                  ) : (
+                    row.peso
+                  )}
+                </td>
+                <td>
+                  {editingRowId === row.id_objeto ? (
+                    <input type="text" defaultValue={row.valor} id={`valor-${row.id_objeto}`} />
+                  ) : (
+                    row.valor
+                  )}
+                </td>
+                <td>
+                  {editingRowId === row.id_objeto ? (
+                    <input type="text" defaultValue={row.cantidad} id={`cantidad-${row.id_objeto}`} />
+                  ) : (
+                    row.cantidad
+                  )}
+                </td>
+                <td>{id}</td>
+                <td>
+                  <button id="boton3" onClick={() => handlerDeleteObjects(row.id_objeto, props.tabla.tipo_inventario)}>Eliminar</button>
+                  {editingRowId === row.id_objeto ? (
+                    <button id="boton1" onClick={() => handleUpdate(row.id_objeto, props.tabla.tipo_inventario)}>Actualizar</button>
+                  ) : (
+                    <button id="boton2"onClick={() => setEditingRowId(row.id_objeto)}>Editar</button>
+                  )}
+                </td>
+              </tr>
+            ))}
               </tbody>
             </table>
             <div id="container2">
@@ -145,7 +227,7 @@ const handlerDeleteObjects = async(idObjeto, tipoInventario)=>{
               <tr>
                 <th>ID objetos personalizados</th>
                 <th>Nombre Objeto</th>
-                <th>Reservación</th>
+                <th>Reservación, Numeración</th>
                 <th>Valor Personalizado</th>
                 <th>ID Inventario</th>
                 <th>Opciones</th>
@@ -155,11 +237,36 @@ const handlerDeleteObjects = async(idObjeto, tipoInventario)=>{
               {updatedObjectsP.map((row,index)=>(
                 <tr key={index}>
                 <td>{row.id_objeto_p}</td>
-                <td>{row.objeto_p}</td>
-                <td>{row.reservacion}</td>
-                <td>{row.valor_p}</td>
+                <td>
+                  {editingRowId === row.id_objeto_p ? (
+                    <input type="text" defaultValue={row.objeto_p} id={`objetoP-${row.id_objeto_p}`} />
+                  ) : (
+                    row.objeto_p
+                  )}
+                </td>
+                <td>
+                  {editingRowId === row.id_objeto_p ? (
+                    <input type="text" defaultValue={row.reservacion} id={`reservacion-${row.id_objeto_p}`} />
+                  ) : (
+                    row.reservacion
+                  )}
+                </td>
+                <td>
+                  {editingRowId === row.id_objeto_p ? (
+                    <input type="text" defaultValue={row.valor_p} id={`valorP-${row.id_objeto_p}`} />
+                  ) : (
+                    row.valor_p
+                  )}
+                </td>
                 <td>{id}</td>
-                <td><button id="boton3" onClick={() =>handlerDeleteObjects(row.id_objeto_p,props.tabla.tipo_inventario)}>Eliminar</button></td>
+                <td>
+                  <button id="boton3" onClick={() => handlerDeleteObjects(row.id_objeto_p, props.tabla.tipo_inventario)}>Eliminar</button>
+                  {editingRowId === row.id_objeto_p ? (
+                    <button id="boton1" onClick={() => handleUpdate(row.id_objeto_p, props.tabla.tipo_inventario)}>Actualizar</button>
+                  ) : (
+                    <button id="boton2"onClick={() => setEditingRowId(row.id_objeto_p)}>Editar</button>
+                  )}
+                </td>
               </tr>
               ))}
               </tbody>
